requirements.txt
kivy
mariadb
python-dotenv
